JAV-564 move EventUtils to utils package
