+++ core/src/main/java/io/servicecomb/core/Invocation.java
@@ -26,7 +26,7 @@ import io.servicecomb.core.definition.SchemaMeta;
 import io.servicecomb.core.metrics.InvocationFinishedEvent;
 import io.servicecomb.core.metrics.InvocationStartProcessingEvent;
 import io.servicecomb.core.provider.consumer.ReferenceConfig;
-import io.servicecomb.foundation.common.event.EventUtils;
+import io.servicecomb.foundation.common.utils.EventUtils;
 import io.servicecomb.swagger.invocation.AsyncResponse;
 import io.servicecomb.swagger.invocation.InvocationType;
 import io.servicecomb.swagger.invocation.SwaggerInvocation;
